+++ shim/shim.js
@@ -5,9 +5,13 @@
 'use strict';
 
 // A shim for the libWrapper library
-export let VERSIONS = [1,7,1];
 export let libWrapper = undefined;
 
+export const VERSIONS       = [1,8,0];
+export const TGT_SPLIT_RE   = new RegExp("([^.[]+|\\[('([^']|\\'|\\\\)+?'|\"([^\"]|\\\"|\\\\)+?\")\\])", 'g');
+export const TGT_CLEANUP_RE = new RegExp("(^\\['|'\\]$|^\\[\"|\"\\]$)", 'g');
+
+// Main shim code
 Hooks.once('init', () => {
 	// Check if the real module is already loaded - if so, use it
 	if(globalThis.libWrapper && !(globalThis.libWrapper.is_fallback ?? true)) {
@@ -22,7 +26,7 @@ Hooks.once('init', () => {
 		static register(package_id, target, fn, type="MIXED", {chain=undefined}={}) {
 			const is_setter = target.endsWith('#set');
 			target = !is_setter ? target : target.slice(0, -4);
-			const split = target.split('.');
+			const split = target.match(TGT_SPLIT_RE).map((x)=>x.replace(/\\(.)/g, '$1').replace(TGT_CLEANUP_RE,''));
 			const fn_name = split.pop();
 			const root_nm = split.splice(0,1)[0];
 			const _eval = eval; // The browser doesn't expose all global variables (e.g. 'Game') inside globalThis, but it does to an eval. We copy it to a variable to have it run in global scope.
@@ -62,11 +66,8 @@ Hooks.once('init', () => {
 	}
 
 	//************** USER CUSTOMIZABLE:
-	// Whether to warn GM that the fallback is being used
-	const WARN_FALLBACK = true;
-
-	// Set up the ready hook that shows the "libWrapper not installed" warning dialog
-	if(WARN_FALLBACK) {
+	// Set up the ready hook that shows the "libWrapper not installed" warning dialog. Remove if undesired.
+	{
 		//************** USER CUSTOMIZABLE:
 		// Package ID & Package Title - by default attempts to auto-detect, but you might want to hardcode your package ID and title here to avoid potential auto-detect issues
 		const [PACKAGE_ID, PACKAGE_TITLE] = (()=>{